# backend/api/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "invoice/v1/invoice.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/products": {
      "get": {
        "operationId": "ProductService_ListProducts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListProductsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ProductService"
        ]
      },
      "post": {
        "operationId": "ProductService_CreateProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateProductRequest"
            }
          }
        ],
        "tags": [
          "ProductService"
        ]
      }
    },
    "/v1/products/{id}": {
      "get": {
        "operationId": "ProductService_GetProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ProductService"
        ]
      },
      "delete": {
        "operationId": "ProductService_DeleteProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ProductService"
        ]
      }
    },
    "/v1/products/{product.id}": {
      "put": {
        "operationId": "ProductService_UpdateProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "product.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductServiceUpdateProductBody"
            }
          }
        ],
        "tags": [
          "ProductService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "ProductServiceUpdateProductBody": {
      "type": "object",
      "properties": {
        "product": {
          "type": "object",
          "properties": {
            "sku": {
              "type": "string"
            },
            "title": {
              "type": "string"
            },
            "price": {
              "type": "string"
            },
            "image": {
              "type": "string"
            },
            "quantity": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CreateProductRequest": {
      "type": "object",
      "properties": {
        "product": {
          "$ref": "#/definitions/v1Product"
        }
      }
    },
    "v1CreateProductResponse": {
      "type": "object",
      "properties": {
        "product": {
          "$ref": "#/definitions/v1Product"
        }
      }
    },
...
    "v1DeleteProductResponse": {
      "type": "object"
    },
...
    "v1GetProductResponse": {
      "type": "object",
      "properties": {
        "product": {
          "$ref": "#/definitions/v1Product"
        }
      }
    },
...
    "v1ListProductsResponse": {
      "type": "object",
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Product"
          }
        }
      }
    },
...
    "v1Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "sku": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "v1UpdateProductResponse": {
      "type": "object",
      "properties": {
        "product": {
          "$ref": "#/definitions/v1Product"
        }
      }
    },
...
  }
}